+++ exinst/default.nix
@@ -1,4 +1,4 @@
-{ mkDerivation, aeson, base, constraints, generic-random
+{ mkDerivation, aeson, base, bytes, constraints, generic-random
 , profunctors, QuickCheck, singletons, stdenv, tasty, tasty-hunit
 , tasty-quickcheck
 }:
@@ -7,10 +7,10 @@ mkDerivation {
   version = "0.2";
   src = ./.;
   libraryHaskellDepends = [
-    aeson base constraints profunctors QuickCheck singletons
+    aeson base bytes constraints profunctors QuickCheck singletons
   ];
   testHaskellDepends = [
-    aeson base constraints generic-random profunctors QuickCheck
+    aeson base bytes constraints generic-random profunctors QuickCheck
     singletons tasty tasty-hunit tasty-quickcheck
   ];
   homepage = "https://github.com/k0001/exinst";